+++ modules/ui_definitions.py
@@ -606,7 +606,7 @@ def create_settings(cmd_opts):
         "huggingface_sep": OptionInfo("<h2>Huggingface</h2>", "", gr.HTML),
         "diffuser_cache_config": OptionInfo(True, "Use cached model config when available"),
         "huggingface_token": OptionInfo('', 'HuggingFace token', gr.Textbox, {"lines": 2}, secret=True, env_var='HF_TOKEN'),
-        "hf_transfer_mode": OptionInfo("rust", "HuggingFace download method", gr.Radio, {"choices": ['requests', 'rust', 'xet']}),
+        "hf_transfer_mode": OptionInfo("XET", "HuggingFace download method", gr.Dropdown, {"choices": ['HTTP', 'XET', 'XET HighPerformance', 'XET Sequential']}),
         "huggingface_mirror": OptionInfo('', 'HuggingFace mirror', gr.Textbox),
         "offline_mode": OptionInfo(False, 'Force offline mode', gr.Checkbox),